reviewed security-utils
