--- common/security-utils/src/main/java/com/mmadu/security/CheckTokenRequest.java
@@ -1,22 +1,0 @@
-package com.mmadu.security;
-
-public class CheckTokenRequest {
-    private String domainId;
-    private String token;
-
-    public String getDomainId() {
-        return domainId;
-    }
-
-    public void setDomainId(String domainId) {
-        this.domainId = domainId;
-    }
-
-    public String getToken() {
-        return token;
-    }
-
-    public void setToken(String token) {
-        this.token = token;
-    }
-}
