+++ common/security-utils/src/test/java/com/mmadu/security/RemoteAppTokenServiceDomainTokenCheckerTest.java
@@ -26,13 +26,18 @@ public class RemoteAppTokenServiceDomainTokenCheckerTest {
     private RemoteAppTokenServiceDomainTokenChecker tokenChecker = new RemoteAppTokenServiceDomainTokenChecker();
 
     @Before
-    public void setUp(){
+    public void setUp() {
         tokenChecker.setTokenServiceUrl("http://localhost:19998");
         tokenChecker.setAdminKey(ADMIN_KEY);
     }
 
     @Test
     public void givenTokenMatchesThenReturnTrue() {
+        mockCheckDomainApiToReturn("{\"matches\": true}");
+        assertThat(tokenChecker.checkIfTokenMatchesDomainToken(TOKEN_ID, DOMAIN_ID), CoreMatchers.is(true));
+    }
+
+    private void mockCheckDomainApiToReturn(String s) {
         stubFor(post(urlPathEqualTo("/token/checkDomainToken"))
                 .withRequestBody(matchingJsonPath(
                         "$.token", WireMock.equalTo(TOKEN_ID))
@@ -42,28 +47,14 @@ public class RemoteAppTokenServiceDomainTokenCheckerTest {
                 )
                 .withHeader("domain-auth-token", WireMock.equalTo(ADMIN_KEY))
                 .willReturn(aResponse()
-                        .withBody("{\"matches\": true}")
+                        .withBody(s)
                         .withHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
                         .withStatus(200)));
-
-        assertThat(tokenChecker.checkIfTokenMatchesDomainToken(TOKEN_ID, DOMAIN_ID), CoreMatchers.is(true));
     }
 
     @Test
     public void givenTokenNotMatchesThenReturnFalse() {
-        stubFor(post(urlPathEqualTo("/token/checkDomainToken"))
-                .withRequestBody(matchingJsonPath(
-                        "$.token", WireMock.equalTo(TOKEN_ID))
-                )
-                .withRequestBody(matchingJsonPath(
-                        "$.domainId", WireMock.equalTo(DOMAIN_ID))
-                )
-                .withHeader("domain-auth-token", WireMock.equalTo(ADMIN_KEY))
-                .willReturn(aResponse()
-                        .withBody("{\"matches\": false}")
-                        .withHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-                        .withStatus(200)));
-
+        mockCheckDomainApiToReturn("{\"matches\": false}");
         assertThat(tokenChecker.checkIfTokenMatchesDomainToken(TOKEN_ID, DOMAIN_ID), CoreMatchers.is(false));
     }
 }
